Fixed speed calc math

+++ src/carInput.hpp
@@ -7,7 +7,7 @@
 // speed calc variables
 const uint16_t samplePeriod = 1000;                                             // how long to get vss samples for, the longer the more accurate
 const uint16_t pulsesPerMile = 4000;                                            // vss pulses per mile, my oldsmobile alero does 4000
-const float convertMph = pulsesPerMile/3600;                                    // 3600 seconds in an hour, since miles per *hour*
+const float convertMph = (float)pulsesPerMile/3600.0;                                    // 3600 seconds in an hour, since miles per *hour*
 static uint8_t roundedMph;
 static uint8_t previousMph;
 static uint8_t speedOut = 0;
